pyproject.toml
[project]
name = "cv-screener-2026"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ollama>=0.6.1",
    "pdfplumber>=0.11.9",
    "python-dotenv>=1.2.1",
    "supabase>=2.27.3",
]
